requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
jinja2==3.1.4
qrcode[pil]==7.4.2
